BugFix:MADLIB-702
Eliminate warnings on gcc 4.6.

There are two kinds of warnings.

First, type conversion, from uint64_t to double, between signed and
unsigned types. For type conversion, we add cast explicitly.

Second, unused auto variables. We just remove them.

+++ src/modules/convex/type/state.hpp
@@ -101,7 +101,7 @@ public:
      * me. But I am not sure where I can move this to...
      */
     inline void computeRMSE() {
-        task.RMSE = sqrt(algo.loss / algo.numRows);
+        task.RMSE = sqrt(algo.loss / static_cast<double>(algo.numRows));
     }
 
     static inline uint32_t arraySize(const uint16_t inRowDim,
