+++ src/modules/regress/LinearRegression_impl.hpp
@@ -42,9 +42,9 @@ void
 LinearRegressionAccumulator<Container>::bind(ByteStream_type& inStream) {
     inStream
         >> numRows >> widthOfX >> y_sum >> y_square_sum;
-    uint16_t actualWidthOfX = widthOfX.isNull()
+    uint16_t actualWidthOfX = static_cast<uint16_t>(widthOfX.isNull()
         ? 0
-        : static_cast<uint16_t>(widthOfX);
+        : static_cast<uint16_t>(widthOfX));
     inStream
         >> X_transp_Y.rebind(actualWidthOfX)
         >> X_transp_X.rebind(actualWidthOfX, actualWidthOfX);
